Početak rada na prikazu serijala. - prikaz podataka o serijalu - akcije na serijalu - sezone na serijalu - epizode po serijalu

+++ lib/models/DefVodContent.dart
@@ -61,7 +61,7 @@ class DefVodContent {
   }
 
   String displayPrice() {
-    if ( int.parse(buyPrice) > 0 ) {
+    if ( double.parse(buyPrice) > 0 ) {
       return buyPrice + '€';
     } else {
       return '';
